rename: 404page style파일 분리
