+++ src/pages/NotFound/style.ts
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const WarningContainer = styled.div`
+  text-align: center;
+  padding: 100px 20px;
+`;
+
+export const WarningStatus = styled.div`
+  font-size: 160px;
+  color: royalblue;
+`;
+
+export const WarningText = styled.div`
+  font-size: 50px;
+`;
